pyproject.toml
[tool.poetry]
name = "iiif-paleography"
version = "0.1.0"
description = ""
authors = ["Mark Baggett <markpbaggett@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-genai = "^1.56.0"
iiif-prezi3 = {git = "https://github.com/markpbaggett/iiif-prezi3.git", rev = "pydantic2-but-still-1"}
click = "^8.3.1"
tqdm = "^4.67.1"
toon-format = {git = "https://github.com/toon-format/toon-python.git"}

[tool.poetry.scripts]
iiif-transcribe = "iiif_paleography.transcribe:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
